Put DB config in a separate object so we can have unimplemented properties (#24)
### Description
Development DB will not have a R/O mode, but all properties are read when `application.config.from_object` is called.  This adds one more layer to the config, just for DB config, where some properties do not need to be implemented.

### Test plan
Loading the app in dev mode works now!