Refactor enumtables hooks
### Description
Enumtables require entries to be put there after the table is created, but there is not a convenient sqlalchemy way to do so.  This dedups the logic that I've been sprinkling all over the codebase to do this work after table creation.  This is now done automatically by `enumtables.EnumTable`.

#### Issue
[ch64743](https://app.clubhouse.io/genepi/stories/space/64743)

### Test plan
Dropped the database and created it from scratch.  Verified that one of the enumtables had the correct values.