Rework the relationship between workflows and entities.
### Description
The key changes is that only one workflow is considered to ever be responsible for creating an entity.  If more than one workflow contributes to an entity, we should have separate entities to represent them.

Added a method to easily traverse between entities and its parents (entities that feed into the workflow that creates the entity in question) and children (entities that result from workflows that consume the entity in question).

### Test plan
used in next PR