Allow for multiple staging environments
### Description
Creates an easy-to-setup staging environment.  This allows developers to create arbitrary staging environments without setting up a new secret.  The callback URL is determined by extracting the Elastic Beanstalk environment name.  Note that this is actually dependent on the cname, but if all the instructions are followed, the cname should match the environment name.

Also updated docs to reflect the new setup.

### Test plan
1. Created and deployed to an AWS Elastic Beanstalk environment.
2. Run locally.