Do not cache the DB interface without the db uri as a cachekey (#128)
### Description
`@lru_cache` takes the arguments of the method as a cachekey.  Since DB URI can change during the course of a unit test run (we change the database name so each test has its own clean database to work with), we end up pulling a stale db handle.  By adding the db_uri as a parameter to the `_DATABASE_INTERFACE` method, we guarantee the correct interface.

### Test plan
Successfully passed other tests.