terraform setup for RDS (#164)
### Description
This creates the database to the point where the user and the role is set up with some default passwords.  To ensure that we do not store the db credentials in the terraform state file, we do not set the password in AWS secrets or manage passwords from terraform.  An additional script is run to pull the secret and set the correct passwords.

#### Issue
[ch<fill_in_issue_number>](https://app.clubhouse.io/genepi/stories/space/<fill_in_issue_number>)

### Test plan
% terraform apply
% aspen-cli db set-passwords-from-secret --environment staging

Then verified I could log into the database.