Install the same versions of the style check tools as in Pipfile.lock
### Description
We are currently always installing the latest version of the style check tools, but this may result in [failures](https://github.com/chanzuckerberg/aspen/runs/2114899925?check_suite_focus=true) when the version we have pinned does not match the behavior of the latest version.  This ensures that developers have a sane experience between their local computers and github acitons.

### Test plan
Github actions.