Re-add dev access to the postgres db
### Description
Setting up the bastion host via terraform and using the bastion host to access the postgres instance via terraform is considerably more effort than we ought to be expending on this.

I added a couple tickets[[1]](https://app.clubhouse.io/genepi/story/126386/add-ssh-bastion-host-that-is-configured-by-terraform)[[2]](https://app.clubhouse.io/genepi/story/126389/existing-tooling-and-terraform-rules-that-access-the-database-should-go-through-the-ssh-bastion-host) to capture this future work.

### Test plan
terraform apply completes