Convert accessions to use a workflow model (#230)
### Description
This turned out to be a lot more complicated than I expected.  Most of the complexity lies in the conversion logic (which we are **not** doing) and the logic for sample view.  The latter is required and now adds another roundtrip to the server.  I suspect it should be possible to do what we want in one query, but I can't seem to massage sqlalchemy to do what I want.

Rest of this is pretty straightforward.

#### Issue
[ch123630](https://app.clubhouse.io/genepi/story/123630/move-accessions-to-a-workflow-model)

### Test plan
make -j style unit-tests