update CZB gisaid dump using the latest processed gisaid dump in aspen (#281)
### Description
This is a stop-gap solution for gisaid ingest for the existing covidtracker infrastructure.  The github action is out of disk space and further investment in that pipeline is not worth it.  This is a small isolated hack that we can do to provide gisaid data to covidtracker.

#### Issue
[ch133367](https://app.clubhouse.io/genepi/story/133367/gisaid-data-ingest-in-biohub-s-infrastructure-is-broken)

### Test plan
`aws batch submit-job --job-name 'update-czb-gisaid' --job-queue aspen-batch --job-definition aspen-batch-job-definition --container-overrides '{"command": ["ttung/czb-gisaid", "src/py/workflows/update_czb_gisaid/update.sh"]}'`