Rename the tree data with private identifiers
### Description
Nodes on trees should be rewritten with the private identifiers that the user is privy to.

#### Issue
[ch133611](https://app.clubhouse.io/genepi/story/133611/rewrite-tree-data-with-private-identifiers)

### Test plan
Added a test with samples from own-group, can-see (private identifiers), can-see (not private identifiers), no can-see, and not even in the system.

Note that I'm stealing the `ncov_aspen.json` file for the `test_auspice_redirect.py` test; fortunately, that test is agnostic of the content of the test.