Update covidhub import to handle internal CZBID (#298)
### Description
This adds support for Internal CZBIDs.

The old workflow was:

1. look at all consensus genomes.
2. look at all CZBIDs that are not represented by the consensus genomes.
3. write out entries for the consensus genomes.
4. write out entries for the failed genomes.

This refactors the workflow to:

1. Pull all the CZBIDs associtaed with the project.
2. Load all the consensus genomes.
3. Loop through all the CZBIDs and:
   a. Find the consensus genome
   b. Sort out the metadata, depending on the CZBID type.

There is an additional wrinkle in how we handle internal czbids, as some samples have multiple CZBIDs.  This can happen if they are sequenced multiple times.  In this case, we pick the CZBID that has the most recovered sites, and if there is a tie there, then we pick the most recently sequenced CZBID.

#### Issue
[ch133318](https://app.clubhouse.io/genepi/story/133318/port-over-internal-samples-from-covidhub-db-into-appropriate-projects-about-500)

### Test plan
import marin's internal czbids (RR089i)