Fix sample count for phylo trees
### Description
Sample counts should reflect what's in the phylo tree.

The previous query was not working because sqlalchemy was creating an implicit join to the PhyloRun table in the inner query.  What we want is a correlated query, where the PhyloRun is coming from the outer query.  Use an alias to explicitly name which PhyloRun we want.

#### Issue
[ch135382](https://app.clubhouse.io/genepi/story/135382/number-of-samples-in-tree-table-view-seems-off)

### Test plan
Added a test to test the phylo_tree endpoint, which verifies the sample counts.