Show current selecting in Genome Recovery
Root cause here was due to the OPTIONS being set in the render of component.
But value checking to see what is selected is done on ===, so the option
objects never had equality because they were new objects each render.
By putting them outside component, they're always the same, and === works.