Revert pangolin workflow export.py
Earlier version of Nextclade workflow directly used the export.py script,
but needed to tweak the ID used in FASTA file, so added a new CLI arg.
That extra CLI arg for pangolin workflow was really just there to help
a different workflow (Nextclade), but that other workflow now has a
specialized exporting script which means the prior tweak to pangolin
is just confusing since it's used nowhere. Reverting back to before that.