On-demand Lineage/QC/Mutations workflow (Nextclade) (#1476)
* Prevent clobbering when volume mounting

* WIP: centralized method for stripped sequences

* Make FASTA for specified samples

* Capture pathogen info during FASTA creation

* Update docs around DB profiling

* Run nextclade using samples pathogen

* Convert output saving to run on PK ids

* Extract pertinent info from nextclade tag

* WIP: Model changes to capture ref data info

* Migration to capture ref data info

* Save ref data info with nextclade results

* Handle finer-grained lineage from CSV

* Handle errors from nextclade run

* Linting

* Ensure all samples are same pathogen

* Name tweak and a dash of docs

* Revert pangolin workflow export.py

Earlier version of Nextclade workflow directly used the export.py script,
but needed to tweak the ID used in FASTA file, so added a new CLI arg.
That extra CLI arg for pangolin workflow was really just there to help
a different workflow (Nextclade), but that other workflow now has a
specialized exporting script which means the prior tweak to pangolin
is just confusing since it's used nowhere. Reverting back to before that.

* WIP: Set up SFN config

* Set up lineage-qc WDLs

* WIP: model changes to enum tools for QC&Mutations

* Migration: Add enum tables for QC & Mutations

This migration depends on the target DB having no data for any of
tables sample_qc_metrics and sample_mutations because of the NOT NULL
constraint on the foreign key to the new enum tables. If our DBs
do have that, need to go back and modify this migration

* Tie records to usage of Nextclade

* Update/fix test to include new work

* Quiet merge conflict around DB migrations

* Migration: Add MPX pathogen and nextclade datasets

* Need absolute path for script call

* Nullable qc_score

* Migration: qc_score nullable

* Invalid results now have NULL qc_score

* WIP: Python handle pathogen slug arg

* Make use of pathogen_slug in calling

* Use innerjoin to ensure genome rather than check

* Linting