Pull mypy version back a bit to 0.981
While still necessary to upgrade -- see prior commit message -- the latest
version of mypy was giving a lot of new errors in our code. Oldest version
that dealt with the bug around checking when numpy was present was 0.981,
using that reduces number of errors in code seen by mypy to around 50% which
feels a lot more manageable given deadlines.

NOTE: This was done by manually tweaking the poetry.lock file, not running
the `poetry update` and re-generating it since that took ~1hr last time...