Workflow for saving aligned pathogen genomes (#1609)
* WIP: General approach to saving APG

* WIP: Update meaning of valid result for Nextclade

* WIP: Stamp APGs with collective run datetime

* WIP: Messaging around fasta saving

* Add biopython (via pip, I apologize) to deps

* two thumbs, loves docs

* Linting

* Test handles new APG work

* Linting

* Update mypy

This update of mypy was specifically to get around a conflict that was
happening due to some weirdness in how it interacted with numpy, which
was installed in another recent commit. See this GitHub Issue:
https://github.com/python/mypy/issues/13627

* Pull mypy version back a bit to 0.981

While still necessary to upgrade -- see prior commit message -- the latest
version of mypy was giving a lot of new errors in our code. Oldest version
that dealt with the bug around checking when numpy was present was 0.981,
using that reduces number of errors in code seen by mypy to around 50% which
feels a lot more manageable given deadlines.

NOTE: This was done by manually tweaking the poetry.lock file, not running
the `poetry update` and re-generating it since that took ~1hr last time...

* Fix mypy stuff up and/or ignore

* Once more unto the breach, my code, once more

* Huh, I guess it doesn't inherit from BE dockerfile